simpletrendbot-multistock/requirements.txt
basebot22-basebot-JustinGuese
pandas
requests
numpy
tqdm